+++ Heuristic.java
@@ -1,3 +1,3 @@
 public interface Heuristic <StateClass extends State, ActionClass extends Action> {
-    Integer Heuristic(StateClass state);
+    Double Heuristic(StateClass state);
 }
